requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
discord==2.3.2
discord.py==2.3.2
fastapi==0.109.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
imgurpython==1.1.7
multidict==6.0.4
outcome==1.3.0.post0
pydantic==2.6.0
pydantic_core==2.16.1
PyJWT==2.8.0
pyotp==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.35.1
trio==0.24.0
trio-websocket==0.11.1
twilio==8.12.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.27.0.post1
wsproto==1.2.0
yarl==1.9.4
